implemented engine basics
The game has a list of things subscribed to it (typically just a
scene). The scene has a list of entities, which have a list of
components that can act on them in different ways. The scene also
has surface for drawing on, and a method to draw on it using a
transform.